requirements.txt
psycopg2>=2.8.6
asyncio>=3.4.3
beautifulsoup4>=4.9.3
lyricsgenius>=2.0.2
Telethon>=1.17.5
tornado>=6.1
requests>=2.25.0
reportlab>=3.5.55
rtl>=0.4.3
telegraph>=1.4.1
PyYAML>=5.3.1
python-telegram-bot>=13.0.0